+++ python/lsst/ts/logging_and_reporting/version.py
@@ -1,3 +1,3 @@
 # Generated by setuptools_scm
 __all__ = ["__version__"]
-__version__ = "0.1.dev49+gb8bd681.d20240913"
+__version__ = "0.1.dev53+g1c878ef.d20240914"
